switch to own SVG of the Swiss cross
Compared to the Adobe Illustrator-exported version from SIX,
this version is defined on a 166px x 166px viewbox,
which are the recommended dimensions with which the cross
should be overlaid on top of QR codes.

This versions visually matches the PNG version from SIX
when rendered onto the same pixel size (166px x 166px).

But, this version scales better!
The SIX version does not render correctly.
My guess is that it might have to do with their
use of floating point coordinates.